pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.6",
    "django-bootstrap5>=24.2",
    "django-filter>=24.2",
    "gunicorn>=21.2.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.1.1",
    "rollbar>=1.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "coverage>=7.0.0",
]

[tool.ruff]
exclude = [
    "migrations",
    "**/settings.py",
]

[tool.ruff.lint]
select = ["E", "F", "W", "C90"]
